Addition of Widescan feature for WAN discovery. This is preliminary early stage feature which uses GPU offloading to mass generate IPv6 address combinations (provided a 4 segment prefix). Each address is sent 3 ICMP packets which tell the reciever how and where to connect. While addresses are being pinged, incoming pings are also being parsed for any identifiable pings with similar connection info.
Several ISPs publicly register prefixes with multiple segments, which can be used to narrow down and fine tune discovery to regions and client types. So we leverage this information to throw a wide net.
ICMP packets tend to be more lightweight than UDP packets carrying JSON serialized data, and ICMP traverse the network layer, so theoretically it should preform better in terms of speed.

(This is very preliminary and subject to change. GPU offloading is hardware intensive and can consume lots of memory very fast.)