- Fine tuned worker/consumer threading. - Enhanced workload balancing and reduced botleneck when sending packets via parallel consumer threads. - Sectioned off some debug messages with DEBUG_EXTENDED preprocessor to prevent message queue flooding but also allow debugging in the future if needed.
TODO: Delete redundant worker override threads (that was experimental but not fruitful), tinker with parallel workload of consumer, and settle on threading in Start function.

This current version works considerably well and I was able to run it continously without too much fuss. Definitely needs improved error handling and workload balancing as previouslt stated, but this is a decent preliminary point.